+++ services/emr/describe_flow.go
@@ -85,15 +85,15 @@ type DescribeFlowResponse struct {
 	*responses.BaseResponse
 	RequestId               string         `json:"RequestId" xml:"RequestId"`
 	Id                      string         `json:"Id" xml:"Id"`
-	GmtCreate               int            `json:"GmtCreate" xml:"GmtCreate"`
-	GmtModified             int            `json:"GmtModified" xml:"GmtModified"`
+	GmtCreate               int64          `json:"GmtCreate" xml:"GmtCreate"`
+	GmtModified             int64          `json:"GmtModified" xml:"GmtModified"`
 	Name                    string         `json:"Name" xml:"Name"`
 	Description             string         `json:"Description" xml:"Description"`
 	Type                    string         `json:"Type" xml:"Type"`
 	Status                  string         `json:"Status" xml:"Status"`
 	Periodic                bool           `json:"Periodic" xml:"Periodic"`
-	StartSchedule           int            `json:"StartSchedule" xml:"StartSchedule"`
-	EndSchedule             int            `json:"EndSchedule" xml:"EndSchedule"`
+	StartSchedule           int64          `json:"StartSchedule" xml:"StartSchedule"`
+	EndSchedule             int64          `json:"EndSchedule" xml:"EndSchedule"`
 	CronExpr                string         `json:"CronExpr" xml:"CronExpr"`
 	CreateCluster           bool           `json:"CreateCluster" xml:"CreateCluster"`
 	ClusterId               string         `json:"ClusterId" xml:"ClusterId"`
