+++ services/emr/describe_scaling_activity.go
@@ -87,8 +87,8 @@ type DescribeScalingActivityResponse struct {
 	*responses.BaseResponse
 	RequestId     string `json:"RequestId" xml:"RequestId"`
 	BizId         string `json:"BizId" xml:"BizId"`
-	StartTime     int    `json:"StartTime" xml:"StartTime"`
-	EndTime       int    `json:"EndTime" xml:"EndTime"`
+	StartTime     int64  `json:"StartTime" xml:"StartTime"`
+	EndTime       int64  `json:"EndTime" xml:"EndTime"`
 	InstanceIds   string `json:"InstanceIds" xml:"InstanceIds"`
 	TotalCapacity int    `json:"TotalCapacity" xml:"TotalCapacity"`
 	Cause         string `json:"Cause" xml:"Cause"`
